fix #7339 chore(project): move normandy into legacy
Because

* Normandy is only used in legacy

This commit

* Moves Normandy into legacy
