+++ app/experimenter/bugzilla/tests/test_tasks.py
@@ -10,7 +10,7 @@ from experimenter.bugzilla import tasks
 from experimenter.bugzilla.tests.mixins import MockBugzillaMixin
 from experimenter.legacy.legacy_experiments.models import Experiment
 from experimenter.legacy.legacy_experiments.tests.factories import ExperimentFactory
-from experimenter.normandy.tests.mixins import MockNormandyMixin
+from experimenter.legacy.normandy.tests.mixins import MockNormandyMixin
 from experimenter.notifications.models import Notification
