+++ app/experimenter/legacy/normandy/__init__.py
@@ -1,0 +1,1 @@
+from experimenter.legacy.normandy.client import *  # noqa
