+++ app/experimenter/legacy/normandy/serializers.py
