+++ app/experimenter/legacy/normandy/tests/mixins.py
@@ -6,7 +6,7 @@ class MockNormandyMixin(object):
         super().setUp()
 
         mock_normandy_requests_get_patcher = mock.patch(
-            "experimenter.normandy.client.requests.get"
+            "experimenter.legacy.normandy.client.requests.get"
         )
         self.mock_normandy_requests_get = mock_normandy_requests_get_patcher.start()
         self.addCleanup(mock_normandy_requests_get_patcher.stop)
@@ -86,7 +86,7 @@ class MockNormandyTasksMixin(object):
         super().setUp()
 
         mock_tasks_set_is_paused_value_patcher = mock.patch(
-            "experimenter.normandy.tasks.set_is_paused_value_task"
+            "experimenter.legacy.normandy.tasks.set_is_paused_value_task"
         )
         self.mock_tasks_set_is_paused_value = (
             mock_tasks_set_is_paused_value_patcher.start()
@@ -94,7 +94,7 @@ class MockNormandyTasksMixin(object):
         self.addCleanup(mock_tasks_set_is_paused_value_patcher.stop)
 
         mock_tasks_add_start_date_comment_patcher = mock.patch(
-            "experimenter.normandy.tasks.add_start_date_comment_task"
+            "experimenter.legacy.normandy.tasks.add_start_date_comment_task"
         )
         self.mock_tasks_add_start_date_comment = (
             mock_tasks_add_start_date_comment_patcher.start()
@@ -103,7 +103,7 @@ class MockNormandyTasksMixin(object):
         self.addCleanup(mock_tasks_add_start_date_comment_patcher.stop)
 
         mock_tasks_comp_experiment_update_res_patcher = mock.patch(
-            "experimenter.normandy.tasks.comp_experiment_update_res_task"
+            "experimenter.legacy.normandy.tasks.comp_experiment_update_res_task"
         )
         self.mock_tasks_comp_experiment_update_res = (
             mock_tasks_comp_experiment_update_res_patcher.start()
