+++ app/experimenter/legacy/normandy/tests/test_tasks.py
@@ -10,9 +10,12 @@ from experimenter.bugzilla.tests.mixins import MockBugzillaMixin
 from experimenter.legacy.legacy_experiments.constants import ExperimentConstants
 from experimenter.legacy.legacy_experiments.models import Experiment, ExperimentEmail
 from experimenter.legacy.legacy_experiments.tests.factories import ExperimentFactory
-from experimenter.normandy import client as normandy
-from experimenter.normandy import tasks
-from experimenter.normandy.tests.mixins import MockNormandyMixin, MockNormandyTasksMixin
+from experimenter.legacy.normandy import client as normandy
+from experimenter.legacy.normandy import tasks
+from experimenter.legacy.normandy.tests.mixins import (
+    MockNormandyMixin,
+    MockNormandyTasksMixin,
+)
 
 
 @override_settings(CELERY_TASK_ALWAYS_EAGER=True)
