--- app/experimenter/normandy/__init__.py
@@ -1,1 +1,0 @@
-from experimenter.normandy.client import *  # noqa
